Move HTTP clients to their own namespace
