+++ scripts/loadtest-scripts/clients/collectionexerciseclient.py
@@ -4,7 +4,7 @@ import os
 import requests
 from dateutil.relativedelta import relativedelta
 
-from clients.httpcodeexception import HTTPCodeException
+from clients.http.httpcodeexception import HTTPCodeException
 
 collection_exercise_url = os.getenv('COLLECTION_EXERCISE_URL', 'http://localhost:8145')
