+++ scripts/loadtest-scripts/clients/http/authenticatedhttpclient.py
@@ -1,7 +1,7 @@
 import requests
 
 
-class HTTPClient:
+class AuthenticatedHTTPClient:
     def __init__(self, username: str, password: str):
         self.username = username
         self.password = password
