+++ scripts/loadtest-scripts/prepare_loadtests.py
@@ -5,9 +5,9 @@ import os
 from clients.actionclient import ActionClient
 from clients.collectionexerciseclient import CollectionExerciseClient
 from clients.collectioninstrumentclient import CollectionInstrumentClient
-from clients.httpclient import HTTPClient
+from clients.http.authenticatedhttpclient import AuthenticatedHTTPClient
 from clients.sampleclient import SampleClient
-from clients.statuscodecheckinghttpclient import StatusCodeCheckingHTTPClient
+from clients.http.statuscodecheckinghttpclient import StatusCodeCheckingHTTPClient
 
 party_url = os.getenv('PARTY_URL', 'http://localhost:8081')
 party_create_respondent_endpoint = os.getenv('PARTY_CREATE_RESPONDENT_ENDPOINT',
@@ -115,7 +115,7 @@ def main():
         return
 
     # There is work in progress which will remove the need for this step
-    http_client = StatusCodeCheckingHTTPClient(HTTPClient(username=username, password=password))
+    http_client = StatusCodeCheckingHTTPClient(AuthenticatedHTTPClient(username=username, password=password))
     action_client = ActionClient(http_client=http_client,
                                  collection_exercise_client=CollectionExerciseClient(username, password),
                                  service_url=action_url)
